feat: list pagination

+++ backend/core-api/src/modules/contacts/graphql/schemas/customer.ts
@@ -2,7 +2,7 @@ export const types = `
 
   type Customer {
     _id: String
-     state: String
+    state: String
     createdAt: Date
     modifiedAt: Date
     avatar: String
@@ -55,10 +55,11 @@ export const types = `
   type CustomersListResponse {
     list: [Customer],
     pageInfo: PageInfo
-    totalCount: Float,
+    totalCount: Int,
   }
 
 `;
+
 export const conformityQueryFields = `
   conformityMainType: String
   conformityMainTypeId: String
